feat: added dto, models, post and get orders endpoints, added order service and order repository. Now user can send and get orders

+++ internal/middleware/auth_middleware.go
@@ -87,7 +87,7 @@ func WithAuth(logger *zap.SugaredLogger) func(next http.Handler) http.Handler {
 	}
 }
 
-func GetUserID(r *http.Request) (int, bool) {
-	id, ok := r.Context().Value(userIDKey).(int)
+func GetUserID(r *http.Request) (int64, bool) {
+	id, ok := r.Context().Value(userIDKey).(int64)
 	return id, ok
 }
